convert Reload.jsx to typescript
