+++ app/actions/game.js
@@ -1,12 +1,9 @@
 import { API } from '../utils/api';
-import { checkVersion } from './utils';
 
 export const SET_GAMES = 'SET_GAMES';
 
 export function listGames () {
   return (dispatch) => {
-    dispatch(checkVersion());
-
     return API.get('/games')
     .then((games) => {
       dispatch(setGames(games));
