+++ app/components/pages/Home.jsx
@@ -2,22 +2,16 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { Link } from 'react-router-dom';
 import { faLongArrowAltRight, faRss } from '@fortawesome/free-solid-svg-icons';
 import { faTwitter } from '@fortawesome/free-brands-svg-icons';
-import { useDispatch } from 'react-redux';
 import { useEffect } from 'react';
 
-import { checkVersion } from '../../actions/utils';
 import { useSession } from '../../hooks/contexts/use-session';
 
 export function Home () {
-  const dispatch = useDispatch();
-
   const { session } = useSession();
 
   useEffect(() => {
     document.title = 'Pokédex Tracker | Track the Progress of Your Living Dex Completion';
   }, []);
-
-  dispatch(checkVersion());
 
   return (
     <div className="home-container">
