+++ app/reducers/index.js
@@ -6,7 +6,6 @@ import { dexTypesById } from './dex-types-by-id';
 import { games } from './games';
 import { gamesById } from './games-by-id';
 import { pokemon } from './pokemon';
-import { reload } from './reload';
 import { showInfo } from './show-info';
 import { users } from './users';
 
@@ -17,7 +16,6 @@ export const reducer = combineReducers({
   games,
   gamesById,
   pokemon,
-  reload,
   showInfo,
   users,
 });
